pi_services/requirements.txt
# ===============================
# Core Computer Vision (Python 3.13 compatible)
# ===============================
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0

# ===============================
# Raspberry Pi Camera
# ===============================
picamera2

# ===============================
# Machine Learning (ONNX only - Python 3.13 compatible)
# ===============================
onnxruntime>=1.16.0

# ===============================
# Database
# ===============================
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# ===============================
# HTTP Requests
# ===============================
requests>=2.31.0

# ===============================
# Speech Recognition and TTS
# ===============================
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
pygame>=2.5.0

# ===============================
# Audio dependencies
# ===============================
pyaudio>=0.2.13
